+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
+### Added
+
+- `alwaysFetchFresh` option to force `fetchEvents` to always request events from Google
+
 ### Changed
 
 - Subscriber is now notified immediately upon subscribing
+- By default `fetchEvents` now only fetches a range of events if it has not already been fetched
 
 ### Fixed